requirements.txt
FastAPI
uvicorn[standard]
SQLAlchemy
pydantic
alembic
psycopg[binary,pool]>=3.2.0
pytest
httpx
python-dotenv
pandas
boto3==1.34.128
botocore==1.34.128
aioboto3
urllib3<2
python-magic
python-multipart
asyncpg
fastapi-cache2
aioredis
pytest-asyncio
aiosqlite